refactor(View): ConsoleView 로 View 코드 이동 for 콘솔 로또, 웹 로또 코드 분리
