+++ src/App/ConsoleApp.js
@@ -3,8 +3,8 @@ import LottoList from "../Model/LottoList.js";
 import Rate from "../Model/Rate.js";
 import { parsingNumbers, stringToNumber } from "../utils/parsing.js";
 import Validator from "../Validator.js";
-import InputView from "../View/InputView.js";
-import OutputView from "../View/OutputView.js";
+import InputView from "../View/ConsoleView/InputView.js";
+import OutputView from "../View/ConsoleView/OutputView.js";
 
 class App {
   #validator;
